Move msgraph client to a new module.
