+++ motor/providers/ms365/client.go
@@ -5,45 +5,41 @@ import (
 	"github.com/Azure/azure-sdk-for-go/sdk/azidentity"
 	"github.com/cockroachdb/errors"
 	a "github.com/microsoft/kiota-authentication-azure-go"
+	"go.mondoo.com/cnquery/motor/providers/ms365/msgraphclient"
 	"go.mondoo.com/cnquery/motor/vault"
 )
 
-const DefaultMSGraphScope = "https://graph.microsoft.com/.default"
-
-var (
-	DefaultMSGraphScopes = []string{DefaultMSGraphScope}
-	DefaultRoles         = []string{
-		"Application.Read.All",
-		"AuditLog.Read.All",
-		"Calendars.Read",
-		"Device.Read.All",
-		"DeviceManagementApps.Read.All",
-		"DeviceManagementConfiguration.Read.All",
-		"DeviceManagementManagedDevices.Read.All",
-		"DeviceManagementRBAC.Read.All",
-		"DeviceManagementServiceConfig.Read.All",
-		"Directory.Read.All",
-		"Domain.Read.All",
-		"IdentityProvider.Read.All",
-		"IdentityRiskEvent.Read.All",
-		"IdentityRiskyUser.Read.All",
-		"InformationProtectionPolicy.Read.All",
-		"MailboxSettings.Read",
-		"Organization.Read.All",
-		"OrgContact.Read.All",
-		"Policy.Read.All",
-		"Policy.Read.ConditionalAccess",
-		"Policy.Read.PermissionGrant",
-		"RoleManagement.Read.All",
-		"SecurityActions.Read.All",
-		"SecurityEvents.Read.All",
-		"TeamsApp.Read.All",
-		"TeamSettings.Read.All",
-		"ThreatAssessment.Read.All",
-		"ThreatIndicators.Read.All",
-		"User.Read.All",
-	}
-)
+var DefaultRoles = []string{
+	"Application.Read.All",
+	"AuditLog.Read.All",
+	"Calendars.Read",
+	"Device.Read.All",
+	"DeviceManagementApps.Read.All",
+	"DeviceManagementConfiguration.Read.All",
+	"DeviceManagementManagedDevices.Read.All",
+	"DeviceManagementRBAC.Read.All",
+	"DeviceManagementServiceConfig.Read.All",
+	"Directory.Read.All",
+	"Domain.Read.All",
+	"IdentityProvider.Read.All",
+	"IdentityRiskEvent.Read.All",
+	"IdentityRiskyUser.Read.All",
+	"InformationProtectionPolicy.Read.All",
+	"MailboxSettings.Read",
+	"Organization.Read.All",
+	"OrgContact.Read.All",
+	"Policy.Read.All",
+	"Policy.Read.ConditionalAccess",
+	"Policy.Read.PermissionGrant",
+	"RoleManagement.Read.All",
+	"SecurityActions.Read.All",
+	"SecurityEvents.Read.All",
+	"TeamsApp.Read.All",
+	"TeamSettings.Read.All",
+	"ThreatAssessment.Read.All",
+	"ThreatIndicators.Read.All",
+	"User.Read.All",
+}
 
 func (p *Provider) Auth() (*a.AzureIdentityAuthenticationProvider, error) {
 	var credential azcore.TokenCredential
@@ -70,5 +66,5 @@ func (p *Provider) Auth() (*a.AzureIdentityAuthenticationProvider, error) {
 		return nil, errors.New("invalid secret configuration for ms365 transport: " + p.cred.Type.String())
 	}
 
-	return a.NewAzureIdentityAuthenticationProviderWithScopes(credential, DefaultMSGraphScopes)
+	return a.NewAzureIdentityAuthenticationProviderWithScopes(credential, msgraphclient.DefaultMSGraphScopes)
 }
