+++ motor/providers/ms365/msgraphconv/structs.go
